refactor json decoder
