+++ engine/access/rest/rest_api.go
@@ -2,7 +2,9 @@ package rest
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
+	"io"
 	"net/http"
 	"strings"
 
@@ -79,6 +81,24 @@ func (restAPI *APIHandler) BlocksIdGet(w http.ResponseWriter, r *http.Request) {
 	restAPI.jsonResponse(w, blocks, errorLogger)
 }
 
+func (restAPI *APIHandler) TransactionIDGet(w http.ResponseWriter, r *http.Request) {
+	errorLogger := restAPI.logger.With().Str("request_url", r.URL.String()).Logger() // todo(sideninja) refactor this to be initialized for us
+
+	vars := mux.Vars(r)
+	id, err := toID(vars["id"])
+	if err != nil {
+		restAPI.errorResponse(w, http.StatusBadRequest, fmt.Sprintf("invalid ID"), errorLogger)
+	}
+
+	tx, err := restAPI.backend.GetTransaction(r.Context(), id)
+	if err != nil {
+		restAPI.errorResponse(w, http.StatusBadRequest, fmt.Sprintf("transaction fetching error: %w", err), errorLogger)
+	}
+
+	res := transactionResponse(tx)
+	restAPI.jsonResponse(w, res, errorLogger)
+}
+
 func (restAPI *APIHandler) jsonResponse(w http.ResponseWriter, responsePayload interface{}, errorLogger zerolog.Logger) {
 	encodedBlocks, err := json.Marshal(responsePayload)
 	if err != nil {
@@ -94,6 +114,57 @@ func (restAPI *APIHandler) jsonResponse(w http.ResponseWriter, responsePayload i
 		return
 	}
 	w.WriteHeader(http.StatusOK)
+}
+
+func (restAPI *APIHandler) jsonDecode(body io.ReadCloser, dst interface{}) error {
+	// validate size
+
+	dec := json.NewDecoder(body)
+	dec.DisallowUnknownFields()
+
+	err := dec.Decode(&dst)
+	if err != nil {
+		var syntaxError *json.SyntaxError
+		var unmarshalTypeError *json.UnmarshalTypeError
+
+		switch {
+		case errors.As(err, &syntaxError):
+			msg := fmt.Sprintf("Request body contains badly-formed JSON (at position %d)", syntaxError.Offset)
+			return &badRequest{status: http.StatusBadRequest, msg: msg}
+
+		case errors.Is(err, io.ErrUnexpectedEOF):
+			msg := fmt.Sprintf("Request body contains badly-formed JSON")
+			return &badRequest{status: http.StatusBadRequest, msg: msg}
+
+		case errors.As(err, &unmarshalTypeError):
+			msg := fmt.Sprintf("Request body contains an invalid value for the %q field (at position %d)", unmarshalTypeError.Field, unmarshalTypeError.Offset)
+			return &badRequest{status: http.StatusBadRequest, msg: msg}
+
+		case strings.HasPrefix(err.Error(), "json: unknown field "):
+			fieldName := strings.TrimPrefix(err.Error(), "json: unknown field ")
+			msg := fmt.Sprintf("Request body contains unknown field %s", fieldName)
+			return &badRequest{status: http.StatusBadRequest, msg: msg}
+
+		case errors.Is(err, io.EOF):
+			msg := "Request body must not be empty"
+			return &badRequest{status: http.StatusBadRequest, msg: msg}
+
+		case err.Error() == "http: request body too large":
+			msg := "Request body must not be larger than 1MB"
+			return &badRequest{status: http.StatusRequestEntityTooLarge, msg: msg}
+
+		default:
+			return err
+		}
+	}
+
+	err = dec.Decode(&struct{}{})
+	if err != io.EOF {
+		msg := "Request body must only contain a single JSON object"
+		return &badRequest{status: http.StatusBadRequest, msg: msg}
+	}
+
+	return nil
 }
 
 func (restAPI *APIHandler) NotImplemented(w http.ResponseWriter, r *http.Request) {
@@ -118,4 +189,13 @@ func (restAPI *APIHandler) errorResponse(w http.ResponseWriter, returnCode int, 
 	if err != nil {
 		logger.Error().Err(err).Msg("failed to send error response")
 	}
+}
+
+type badRequest struct {
+	status int
+	msg    string
+}
+
+func (mr *badRequest) Error() string {
+	return mr.msg
 }
